+++ extra/files/gen2-centos/provision.sh
@@ -123,8 +123,9 @@ timedatectl set-timezone Europe/Copenhagen --no-ask-password
 echo "Provisioning phase 3 - Nefoetch"
 curl -o /etc/yum.repos.d/konimex-neofetch.repo https://copr.fedorainfracloud.org/coprs/konimex/neofetch/repo/epel-7/konimex-neofetch-epel-7.repo
 echo "Provisioning phase 3 - Extra Packages"
-yum -y install htop atop iftop iotop firewalld bmon nmap realmd samba nmon samba-common oddjob oddjob-mkhomedir sssd ntpdate ntp adcli krb5-workstation sssd-libwbclient jq firefox gparted pv neofetch screen telnet ncdu tmux multitail
-
+yum -y install htop atop iftop iotop firewalld bmon nmap realmd samba nmon samba-common oddjob oddjob-mkhomedir sssd ntpdate ntp adcli krb5-workstation sssd-libwbclient jq firefox gparted pv neofetch screen telnet ncdu tmux multitail rkhunter
+echo "Provisioning phase 3 - RK hunter"
+rkhunter --propupd
 echo "Provisioning phase 3 - MOTD"
 
 if [ -f /tmp/motd.sh ]; then