+++ hv_centos78_docker.ps1
@@ -4,9 +4,9 @@
 $startDTM = (Get-Date)
 
 # Variables
-$template_file="./templates/hv_centos7_g2.json"
-$var_file="./variables/variables_centos77.json"
-$machine="CentOS 7.7 1908"
+$template_file="./templates/hv_centos7_g2_docker.json"
+$var_file="./variables/variables_centos78.json"
+$machine="CentOS 7.8 2003 Docker"
 $packer_log=0
 
 if ((Test-Path -Path "$template_file") -and (Test-Path -Path "$var_file")) {